requirements.txt
Flask==3.0.0
python-telegram-bot==20.5
ccxt==4.2.85
pandas==2.1.4
numpy==1.26.0
scikit-learn==1.3.2
matplotlib==3.8.0
joblib==1.3.2
python-dotenv==1.0.0